requirements/req.txt
Flask==3.1.0
colorama==0.4.6     
python-dotenv==1.0.1
ytmusicapi==1.10.2  
ntplib==0.4.0       
pytz==2025.1        
yt-dlp==2025.2.19
bcrypt==4.3.0
mutagen==1.47.0
pyyaml==6.0.2
requests==2.32.3
pyfiglet==1.0.2
gunicorn==23.0.0
tqdm==4.67.1
termcolor==2.5.0
bs4==0.0.2